requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.31
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
aiosqlite==0.20.0
psycopg2-binary==2.9.9
